+++ P1/src/structs/indice.c
@@ -1,7 +1,7 @@
 #include "indice.h"
 #include <stdlib.h>
 
-Indice* indice_init(unsigned long int tamano, unsigned long int identificador_relativo, unsigned long int identificador_absoluto) {
+Indice* indice_init(unsigned int tamano, unsigned int identificador_relativo, unsigned int identificador_absoluto) {
     Indice* indice = malloc(sizeof(Indice));
     indice -> tamano = tamano;
     indice -> cantidad_bloques = tamano / 2048;
@@ -12,7 +12,7 @@ Indice* indice_init(unsigned long int tamano, unsigned long int identificador_re
 };
 
 void assing_data_list(Indice* indice, Data* data, int index) {
-    // int cantidad_bloques = sizeof data_list_pointer / sizeof unsigned long int;
+    // int cantidad_bloques = sizeof data_list_pointer / sizeof unsigned int;
     indice -> lista_de_datos[index] = data;
 }
