+++ P1/src/structs/indice.h
@@ -2,17 +2,17 @@
 #include "data.h"
 
 typedef struct indice {
-    unsigned long int tamano; // tamano archivo
+    unsigned int tamano; // tamano archivo
     int cantidad_bloques; // tamaño dividido por 2kb
     Data** lista_de_datos; // lista con numeros de bloques
-    unsigned long int identificador_relativo;
-    unsigned long int identificador_absoluto;
+    unsigned int identificador_relativo;
+    unsigned int identificador_absoluto;
 } Indice;
 
 
 // FUNCTIONS
 
-Indice* indice_init(unsigned long int tamano, unsigned long int identificador_relativo, unsigned long int identificador_absoluto);
+Indice* indice_init(unsigned int tamano, unsigned int identificador_relativo, unsigned int identificador_absoluto);
 
 void assing_data_list(Indice* indice, Data* data, int index);
